+++ src/screens/journaling/JournalCategory.js
@@ -1,66 +1,11 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import { Pressable, SafeAreaView, Text, View } from 'react-native';
+import { SafeAreaView, Text, View } from 'react-native';
 import { sizes } from '../../data/theme';
 import { useJournal } from '../../context/Context';
 import PrimaryButton from '../../components/PrimaryButton';
+import EmotionCategoryButton from '../../components/journaling/EmotionCategoryButton';
 
-/**
- * A button component for emotion categories.
- *
- * @component
- * @param {object} props - The component's properties.
- * @param {string} props.title - The title of the emotion category.
- * @returns {JSX.Element} The rendered EmotionCategoryButton component.
- */
-function EmotionCategoryButton({ title }) {
-    const { journal, dispatchJournal } = useJournal();
-    const isSelected = journal.emotionCategory === title;
-
-    /**
-     * Handles the press event of an emotion category button.
-     *
-     * @param {string} title - The title of the emotion category.
-     */
-    const handleEmotionCategoryPress = (title) => {
-        dispatchJournal({ type: 'setJournal', payload: { emotionCategory: title } });
-    };
-
-    return (
-        <>
-            <Pressable
-                onPress={() => handleEmotionCategoryPress(title)}
-                style={{
-                    backgroundColor: isSelected ? 'green' : 'grey',
-                    aspectRatio: 1,
-                    alignItems: 'center',
-                    justifyContent: 'center',
-                    padding: sizes.padding.sm,
-                    width: '40%',
-                    borderRadius: sizes.radius,
-                }}
-            >
-                <Text style={{ fontWeight: 'bold', textAlign: 'center' }}>{title}</Text>
-            </Pressable>
-        </>
-    );
-}
-
-/**
- * Prop types for the EmotionCategoryButton component.
- *
- * @typedef {object} EmotionCategoryButtonProps
- * @property {string} title - The title of the emotion category.
- */
-
-/**
- * Prop types for the EmotionCategoryButton component.
- *
- * @type {EmotionCategoryButtonProps}
- */
-EmotionCategoryButton.propTypes = {
-    title: PropTypes.string.isRequired,
-};
 
 /**
  * A component for selecting the emotion category.
@@ -76,7 +21,7 @@ function JournalCategory({ navigation }) {
 
     return (
         <SafeAreaView>
-            <View style={{ alignItems: 'center', marginHorizontal: sizes.padding.md, gap: sizes.padding.md }}>
+            <View style={{ alignItems: 'center', paddingTop: sizes.padding.md, marginHorizontal: sizes.padding.md, gap: sizes.padding.md }}>
                 <Text style={{ fontSize: sizes.text.header1 }}>Choose How You're Feeling Right Now</Text>
                 <View style={{ width: '100%', flexDirection: 'column', justifyContent: 'space-around', alignItems: 'center' }}>
                     <View style={{ flexDirection: 'row', justifyContent: 'space-around', width: '100%' }}>
